Add links and disable NaN
Link to show average total price is shown on index, while hiding it is
shown on the average_total_price page.

When showing the average price, sometimes employees don't have an order,
and an NaN is shown. Add a check to make sure orders isn't empty so that
the result will be blank.